WinGui: Adding Del shortcut to preset pane, and better warnings when deleting.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@7103 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ win/CS/HandBrakeWPF/ViewModels/MainViewModel.cs
@@ -1638,6 +1638,29 @@ namespace HandBrakeWPF.ViewModels
         {
             if (this.selectedPreset != null)
             {
+                if (this.selectedPreset.IsDefault)
+                {
+                    this.errorService.ShowMessageBox(
+                      "You can not delete the default preset. Please set another preset as default first.",
+                      Resources.Warning,
+                      MessageBoxButton.OK,
+                      MessageBoxImage.Information);
+
+                    return;
+                }
+
+                MessageBoxResult result =
+                this.errorService.ShowMessageBox(
+                   "Are you sure you want to delete the preset: " + this.selectedPreset.Name + " ?",
+                   Resources.Question,
+                   MessageBoxButton.YesNo,
+                   MessageBoxImage.Question);
+
+                if (result == MessageBoxResult.No)
+                {
+                    return;
+                }
+
                 this.presetService.Remove(this.selectedPreset);
             }
             else
